Added Login Style and Sign up Style. Fixed all the submit without checking empty fields
