+++ app/build.gradle
@@ -42,5 +42,6 @@ dependencies {
     annotationProcessor 'com.github.bumptech.glide:compiler:4.11.0'
     //location
     implementation 'com.google.android.gms:play-services-location:17.0.0'
+    // material
     implementation 'com.google.android.material:material:1.1.0'
 }
