--- app/src/main/java/com/rizeup/Login/LoginActivity.java
@@ -1,70 +1,0 @@
-package com.rizeup.Login;
-
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
-
-import com.google.android.gms.tasks.OnCompleteListener;
-import com.google.android.gms.tasks.Task;
-import com.google.firebase.auth.AuthResult;
-import com.google.firebase.auth.FirebaseAuth;
-import com.rizeup.MainMenu.MainMenu;
-import com.rizeup.R;
-import com.rizeup.SignUp.SignUpActivity;
-
-public class LoginActivity extends AppCompatActivity {
-
-    private FirebaseAuth mAuth;
-    private Button loginBtn, signUpBtn;
-    private EditText email, password;
-    private Intent signUp, homePage;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_login);
-        mAuth = FirebaseAuth.getInstance();
-        loginBtn = (Button) findViewById(R.id.btn_login);
-        signUpBtn = (Button) findViewById(R.id.btn_signup);
-        email = (EditText) findViewById(R.id.et_email);
-        password = (EditText) findViewById(R.id.et_password);
-        homePage = new Intent(this, MainMenu.class);
-        signUp = new Intent(this, SignUpActivity.class);
-        loginBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                loginUserAccount();
-            }
-        });
-        signUpBtn.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                startActivity(signUp);
-                finish();
-            }
-        });
-    }
-
-    private void loginUserAccount() {
-        mAuth.signInWithEmailAndPassword(email.getText().toString(), password.getText().toString())
-                .addOnCompleteListener(new OnCompleteListener<AuthResult>() {
-                    @Override
-                    public void onComplete(@NonNull Task<AuthResult> task) {
-                        if (task.isSuccessful()) {
-                            Toast.makeText(getApplicationContext(), "Login successful!", Toast.LENGTH_LONG).show();
-                            startActivity(homePage);
-                        } else {
-                            Toast.makeText(getApplicationContext(), "Login Fail!", Toast.LENGTH_LONG).show();
-                        }
-                    }
-                });
-    }
-}
-
