+++ app/src/main/res/layout/activity_main_menu.xml
@@ -38,6 +38,14 @@
             android:layout_margin="10dp"
             android:text="Manage Queue" />
 
+        <Button
+            android:id="@+id/logoutBtn"
+            android:layout_width="0dp"
+            android:layout_weight="1"
+            android:layout_height="match_parent"
+            android:layout_margin="10dp"
+            android:text="Logout" />
+
     </LinearLayout>
 
     <LinearLayout
